+++ modules/automation-account/automationaccount.tf
@@ -6,6 +6,10 @@ resource "azurerm_automation_account" "automation_account" {
 
   sku_name = var.automation_account_sku
   tags     = merge(local.default_tags, var.extra_tags, var.automation_account_extra_tags)
+
+  identity {
+    type = var.automation_account_identity_type
+  }
 }
 
 # Link automation account to log analytics workspace