+++ modules/automation-account/versions.tf
@@ -3,7 +3,7 @@ terraform {
   required_providers {
     azurerm = {
       source  = "hashicorp/azurerm"
-      version = ">= 2.57"
+      version = ">= 2.96.0"
     }
     azurecaf = {
       source  = "aztfmod/azurecaf"