+++ versions.tf
@@ -4,7 +4,7 @@ terraform {
     # tflint-ignore: terraform_unused_required_providers
     azurerm = {
       source  = "hashicorp/azurerm"
-      version = ">= 2.57"
+      version = ">= 2.96.0"
     }
     # tflint-ignore: terraform_unused_required_providers
     azurecaf = {